customization files explained in the eBook
Ever J. Barbero
ASIN: B0190O0OIY
The objective of this eBook is to show you how to convert LaTeX to ePub, then to MOBI and iBook, using free software, while maintaining all the beautiful features of LaTeX.
You just have to use LaTeX to produce the content as you would normally do for a paper book, but following a few rules to make for an easy transition to ePub. Then, we will use plasTeX to convert LaTeX to XHTML. Finally, we use Sigil to quickly assemble a single ePub file that contains your book.
Absolutely no programming required. You just have to install some software and move some files around. I explain how to install the software and how to use it step by step.
You will be able to produce eBooks with the features that you see in this eBook and in Multifunctional Composites without any programming whatsoever.
We use free, open source software, complemented by a few customizations that I did, which are available from this site. I explain how to install these customizations together with the software they modify, step by step.
It is not necessary to learn how these customizations were programmed, but if you are interested in further customizing the conversion process, you will find an introduction on how to do it and some examples that work (Chapter 5).